Improved CSV and SHP ingest
